job-portal-server/job-portal-article-service/requirements.txt
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
dataclasses==0.8
dnspython==2.1.0
dotenv-python==0.0.1
dynaconf==3.1.5
fastapi==0.68.1
greenlet==1.1.1
grpcio==1.40.0
grpcio-tools==1.40.0
h11==0.12.0
idna==3.2
ifaddr==0.1.7
importlib-metadata==4.8.1
pika==1.2.0
protobuf==3.18.0
psycopg2==2.9.1
py-eureka-client==0.10.0
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
zipp==3.5.0
